requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
python-jose
passlib[bcrypt]
pydantic
